updated load balancer interface
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@2406 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/parallel/datadriven/basis/linear/noboundary/operation/OperationMultipleEvalIterativeSPOCLLinear.cpp
@@ -84,7 +84,7 @@ double OperationMultipleEvalIterativeSPOCLLinear::multTransposeVectorized(sg::ba
 			{
 				float eval = ((ptrLevel[(j*dims)+d]) * (ptrData[(i*dims)+d]));
 				float index_calc = eval - (ptrIndex[(j*dims)+d]);
-				float abs = fabs(index_calc);
+				float abs = (float)fabs(index_calc);
 				float last = 1.0f - abs;
 				float localSupport = std::max<float>(last, 0.0f);
 				curSupport *= localSupport;
